Update ouroboros-consensus, cardano-api, cardano-cli and cardano-rpc for 11.2 release.
* Replace leiosKernelTracer/leiosPeerTracer (removed) with new Peras
  diffusion tracers: perasCertDiffusionInbound/Outbound,
  perasVoteDiffusionInbound/Outbound, perasCertInclusion,
  perasVoteForging — all stubbed with nullTracer for now
* Remove ObjectDiffusion inbound/outbound helpers from Consensus.hs
  (TraceObjectDiffusionInbound/Outbound no longer exported)
* ExtValidationError has two new constructors:
  ExtValidationErrorPerasEpochContextResolver and
  ExtValidationErrorPerasCertInBlock; add pattern matches and
  Show (PerasError blk) constraint to all affected LogFormatting
  instances (TraceEvent, TraceAddBlockEvent, TraceValidationEvent,
  TraceInitChainSelEvent, TraceForgeEvent, ExtValidationError)
* PerasVoteDB.TraceEvent now requires Show (PerasCert blk) instead of
  StandardHash blk for its LogFormatting instance
* Drop stale tracing namespace registrations for removed tracers in
  Consistency.hs and Documentation.hs
* Replace deprecated caseByronOrShelleyBasedEra with inEonForEra
  in tx-generator
* mkNodeKernelAccess gains a ShelleyGenesisFile In argument between the
  genesis hash and block type; extract it alongside the hash in Run.hs
* Update TraceRpc constructor names in Rpc.hs
* Fix incomplete pattern match in namespaceFor for TraceRpcSync:
  add TraceRpcFetchBlockNotFound and TraceRpcNodeKernelAccessUnavailable
* Rename StakeKeyAlreadyRegisteredDELEG to DelegAccountAlreadyRegistered
* Add allow-newer for plutus-scripts-bench plutus-* bounds
* Add PlutusV4 cost model values in emptyDijkstraGenesis (Dijkstra.hs)

+++ bench/tx-generator/src/Cardano/TxGenerator/ProtocolParameters.hs
@@ -326,7 +326,7 @@ requireParam paramName = maybe (Left $ PpceMissingParameter paramName)
 mkProtVer :: (Natural, Word32) -> Either ProtocolParametersConversionError Ledger.ProtVer
 mkProtVer (majorProtVer, minorProtVer) =
   maybeToRight (PpceVersionInvalid majorProtVer) $
-    (`Ledger.ProtVer` minorProtVer) <$> Ledger.mkVersion majorProtVer
+    (`Ledger.ProtVer` fromIntegral minorProtVer) <$> Ledger.mkVersion majorProtVer
 
 -- Duplicated from "cardano-api" module "Cardano.Api.Internal.ProtocolParameters"
 boundRationalEither
@@ -528,7 +528,7 @@ fromShelleyCommonPParams
 fromShelleyCommonPParams pp =
   ProtocolParameters
     { protocolParamProtocolVersion = case pp ^. ppProtocolVersionL of
-        Ledger.ProtVer a b -> (Ledger.getVersion a, b)
+        Ledger.ProtVer a b -> (Ledger.getVersion a, fromIntegral b)
     , protocolParamMaxBlockHeaderSize = fromIntegral $ pp ^. ppMaxBHSizeL
     , protocolParamMaxBlockBodySize = fromIntegral $ pp ^. ppMaxBBSizeL
     , protocolParamMaxTxSize = fromIntegral $ pp ^. ppMaxTxSizeL
